Random fixes/extensions
some random small fixes/extensions to the imaging, pca, and spectralcollection modules

+++ src/pyir_image.py
@@ -661,5 +661,4 @@ class PyIR_Image:
                 rebuild_knn[i] = clusters[count]
                 count = count+1
                 
-        return rebuild_knn.astype(int)
-        
+        return rebuild_knn
